Make all do-check-* functions have same result type.
One of the do-check-* functions was returning the <result> object
while the others were all just returning the <result-status> object.
Making them all return the same type of result will make it easier
to combine the bulk of the logic into a single helper method or
macro. It also makes it easier to extend the testing in the future
to cover the reasons that are generated for test failures.

* assertions.dylan
  (do-check-instance?): Update return values.
  (do-check-true): Likewise.
  (do-check-false): Likewise.
  (do-check-condition): Likewise.

* tests/testworks-test-suite.dylan
  (test testworks-check-true-test,
   test testworks-assert-true-test,
   test testworks-check-false-test,
   test testworks-assert-false-test,
   test testworks-assert-equal-test,
   test testworks-assert-not-equal-test,
   test testworks-check-instance?-test,
   test testworks-assert-instance?-test,
   test testworks-not-assert-instance?-test,
   test testworks-check-condition-test,
   test testworks-assert-signals-test,
   test testworks-check-no-errors-test,
   test testworks-assert-no-errors-test): Update for new return values
    and to get the actual result status from the result object.