Make run-test-application() run all defined tests, without need of suites.
If a top-level suite is specified then it is used, otherwise all root
components (those not contained in any suite) are run. Testworks' test
suite itself is the first test of this. The result is exactly the same
since there's exactly one root suite: testworks-test-suite.

This gets rid of root-suite() and instead registers all components in
a single vector.  This simplifies a bunch of code that finds
components by name and also facilitates finding the root components.