Allow format strings in assertion descriptions
This makes it easier to specify descriptions (a.k.a. names) in the `assert-*`
macros. For example, instead of

  `assert-true(f(i), format-to-string("f(%d)", i))`

we can now write

  `assert-true(f(i), "f(%d)", i)`

or even

  `assert-true(f(i), i)`

if the test is very short and the only thing needed to identify the assertion
is the loop variable i.

As a side-effect this fixes #143.