Add deps as submodules, plus registry files
Until dylan-tool has separate deps and test-deps settings testworks can't be
tested with dylan-tool so instead add submodules. This way testworks is not
tied to the versions of those libraries that are submoduled into whatever Open
Dylan release is being used.  (Specifically, testworks is not compatible with
the version of command-line-parser in OD 2020.1.)