Change action install-dylan-tool for install open dylan
The action install-dylan-tool@v2 uses the version 0.8.0 that creates a
different directory structure for packages. The last version 0.10.1 is
needed and because Opendylan version 2023.1 includes it use this
install-opendylan@v3 and not install-dylan-tool@v2.

Additionaly this action is faster, since install-dylan-tool@v2
installs the compiler and compiles dylan-tool, but
install-opendylan@v3 only uncompress the compiler with the tool in it.