Simplify the summary report
Example:
```
Ran 1276 assertions
Ran 53 tests and 2 benchmarks: 4 expected failures
PASSED in 0.023699 seconds
```

Notably, no attempt is made to report the number of passing/failing assertions;
only tests and benchmarks. The total number of assertions is still displayed
since it can (sometimes) give a signal as to the thoroughness of the tests.