Remove duplicate --progress output
Fixes #180

Output now looks like this:

```
_build/bin/testworks-test-suite-app --suite testworks-results-suite
Suite testworks-results-suite:
  Test test-run-tests/suite: PASSED in 0.003218s and 481KiB
  Test test-run-tests/test: PASSED in 0.000052s and 22KiB
  Test test-run-tests-expect-failure/suite: PASSED in 0.000105s and 18KiB
  Test test-run-tests-expect-failure/test: PASSED in 0.000083s and 21KiB
Ran 13 assertions
Ran 4 tests
PASSED in 0.003458 seconds
```

To simplify the --progress code a bit I decided to only output the assertions once the
entire test is done. Otherwise it's impossible to put the test result and timings on the
same line as the test name. This means that progress is essentially reported with test
granularity now, rather than assertion granularity.