Run suite setup/cleanup when --test is used
This changes the way the runner works a little bit. Instead of checking tags and looking
for suite setup/cleanup during the test run everything is computed in advance and the
exact set of components that will be executed based on command-line args is stored in the
runner-components slot. If a suite is in that set, then its setup/cleanup need to be run.
This also meant that runner-tags and runner-skip were no longer needed.

Fixes #148