Capture standard out/err in the test temp directory
Fixes #83

I'm claiming that this fixes #83 even though it doesn't connect the output with the test
result objects.  This provides a good base that makes that easy to accomplish if we want
to do that later. The main thing is that the output no longer spams the console and yet
it is available for later debugging if needed.