Better temp directories
Adds a `.1`, `.2`, etc uniquifier to the end of the test run directory name so that if
two test runs are executed within the same second there won't be a clash.

The unique directory for the entire test run is now stored in the runner. Previously it
was possible to end up with one test's temp files in `_test/cgay-20250926-101601` and the
next test's temp files in `_test/cgay-20250926-101602`.

Fixes #149
Fixes #199